Remove managed beans subsystem
