--- managedbeans/src/main/resources/org/jboss/as/managedbean/LocalDescriptions.properties
@@ -1,2 +1,0 @@
-managedbeans=The configuration of the managed beans subsystem.
-managedbeans.add=Adds the managed beans subsystem.
